Styled brand card content into a grid type. With the edit card turning blue on hover and delete card turning red on hover.
Automated css to detect the number of navbar items and create a grid accordingly; done using a Javascript function that counts the number of navbar links.

Changed css to use a global color theme. This was done using a css variable (--name) so that all of the website colors can be controlled by changing one variable.